Hairy Messy Ugly shell script to emulate xxd
On client machines where there is no `xxd` binary present, but we do
find `bc` we can now properly un-hexdump the files being sent by the
server.  It's crazy slow for large files but it works well enough.

I also changed the way the push process works, removing the hexdump
transformation when initially sending across the netskel client itself,
instead receiving it via a 'rawclient' server method.  This isn't
awesome, but it's safe enough for a first push and if it fails nothing
automated will be broken.

Resolves #5